Split tests into conftest and ep-merge 'process' class
